HSU: reset TX buffer after DMA
The TX buffer is circular, but DMA wants linear. At the the of the buffer a
message will be split into two transactions, causing an interchar gap on transmission.
The receiving device (especially if another Edison) will interpret the gap as
the end of the message, terminating the receive DMA transaction and setting up
for the next. In the mean time the receive FIFO will overflow and loose characters.

To prevent this when the TX buffer is empty we reset head and tail to the beginning
of the buffer.

This is a simple hack that seems to work well enough. But it might be reverted
in the future, and create a seperate linear TX DMA buffer and copy data into
it from the serial buffer.

Signed-off-by: ferry <ftoth@exalondelft.nl>